+++ src/context/user.context.jsx
@@ -1,18 +1,18 @@
-import { createContext, useEffect, useReducer } from "react";
-import {
-  createUserDocFromAuth,
-  onAuthStateChangedListener,
-} from "../utils/firebase/firebase.utils.js";
-import { createAction } from "../utils/reducer/reducer.utils.js";
+import { createContext, useReducer } from "react";
+// import { createAction } from "../utils/reducer/reducer.utils.js";
 
 export const UserContext = createContext({
   currentUser: null,
   setCurrentUser: () => null,
 });
 
-export const USER_ACTION_TYPES = {
-  SET_CURRENT_USER: "SET_CURRENT_USER",
+const INITIAL_STATE = {
+  currentUser: null,
 };
+
+// export const USER_ACTION_TYPES = {
+//   SET_CURRENT_USER: "SET_CURRENT_USER",
+// };
 
 const userReducer = (state, action) => {
   const { type, payload } = action;
@@ -29,34 +29,30 @@ const userReducer = (state, action) => {
   }
 };
 
-const INITIAL_STATE = {
-  currentUser: null,
-};
-
 export const UserProvider = ({ children }) => {
   const [state, dispatch] = useReducer(userReducer, INITIAL_STATE);
 
   const { currentUser } = state;
 
-  const setCurrentUser = (user) => {
-    dispatch(createAction(USER_ACTION_TYPES.SET_CURRENT_USER, user));
-  };
+  // const setCurrentUser = (user) => {
+  //   dispatch(createAction(USER_ACTION_TYPES.SET_CURRENT_USER, user));
+  // };
 
-  const value = { currentUser, setCurrentUser };
+  // const value = { currentUser, setCurrentUser };
 
-  useEffect(() => {
-    const unsubscribe = onAuthStateChangedListener(async (user) => {
-      if (user) {
-        try {
-          await createUserDocFromAuth(user);
-        } catch (err) {
-          console.log(err);
-        }
-      }
-      setCurrentUser(user);
-    });
-    return unsubscribe;
-  }, []);
+  // useEffect(() => {
+  //   const unsubscribe = onAuthStateChangedListener(async (user) => {
+  //     if (user) {
+  //       try {
+  //         await createUserDocFromAuth(user);
+  //       } catch (err) {
+  //         console.log(err);
+  //       }
+  //     }
+  //     setCurrentUser(user);
+  //   });
+  //   return unsubscribe;
+  // }, []);
 
-  return <UserContext.Provider value={value}>{children}</UserContext.Provider>;
+  // return <UserContext.Provider value={value}>{children}</UserContext.Provider>;
 };
