+++ src/store/user/user.reducer.js
@@ -1,0 +1,18 @@
+const INITIAL_STATE = {
+  currentUser: null,
+};
+
+export const userReducer = (state = INITIAL_STATE, action) => {
+  const { type, payload } = action;
+
+  switch (type) {
+    case "SET_CURRENT_USER":
+      return {
+        ...state,
+        currentUser: payload,
+      };
+
+    default:
+      return state;
+  }
+};
